fix problem of p-value
We fixed a problem identified by the reviewer.
Reviewer comment: "There appears to be an error in the function getinfo() in file fun_trends.R (github, version c7f9966, line 133). If opt2 is TRUE, this function calculates the p-value of the quantile regression by taking the slope coefficient’s t-value and plugging it into the probability density function of the chi-squared distribution. This does not seem the right way to calculate the p-value (for one, t-values can be negative, but the chi-squared distribution cannot, resulting in an artificial p-value of 0 (i.e. a significant trend) whenever the intercept is negative. If this code was used to create the results for the paper, this could have inflated the estimated number of trends. 
The simplest way to solve this issue is to use the p-value provided with the slope coefficient instead (unless there is a strong theoretical reason unknown to me not to trust the t-test in this situation)."